+++ accounts/urls.py
@@ -1,9 +1,11 @@
 from django.urls import path
 from . import views
 
+app_name = 'accounts'
+
 urlpatterns = [
-    path('test/admin/', views.admin_test_view, name='admin_test_view'),
-    path('test/org_owner/', views.org_owner_test_view, name='org_owner_test_view'),
-    path('test/donor/', views.donor_test_view, name='donor_test_view'),
-    path('test/admin_or_org_owner/', views.admin_or_org_owner_test_view, name='admin_or_org_owner_test_view'),
+    path('login/', views.CustomLoginView.as_view(), name='login'),
+    path('logout/', views.CustomLogoutView.as_view(), name='logout'),
+    path('register/donor/', views.DonorRegistrationView.as_view(), name='register_donor'),
+    path('register/org/', views.OrgRegistrationView.as_view(), name='register_org'),
 ]